[리팩토링] 데이터 처리 로직 UI 에서 분리
