+++ src/system/events.rs
@@ -1,7 +1,6 @@
 use bevy::prelude::*;
 
 use crate::common::{status_type::ViewStatusType, direction::Direction};
-use serde::{Deserialize, Serialize};
 
 #[derive(Event)]
 pub struct ChangeGameStatus(pub ViewStatusType);
@@ -36,12 +35,6 @@ pub struct TextPopupEvent {
 pub enum ScoreBoardType {
     Current,
     HighScore,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ScoreData {
-    pub score: u32,
-    pub date: String,
 }
 
 #[derive(Debug, Event)]
