+++ src/system/mod.rs
@@ -3,3 +3,4 @@ pub mod events;
 pub mod handle_keyboard_input;
 pub mod game;
 pub mod resource;
+pub mod data;
