+++ src/ui/show_menu.rs
@@ -47,25 +47,3 @@ pub fn show_menu(
         }
     }
 }
-
-fn spawn_main_menu(
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<ColorMaterial>>
-) {
-    println!("Main Menu Selected!");
-    let mesh = Mesh::from(Rectangle::new(MAIN_MENU_WIDTH, MAIN_MENU_HEIGHT));
-    let material = ColorMaterial::from(Color::rgb(0., 1., 0.));
-
-    let mesh_handle = meshes.add(mesh);
-    let material_handle = materials.add(material);
-
-    commands.spawn((
-        MainMenuHandle,
-        MaterialMesh2dBundle {
-            mesh: mesh_handle.into(),
-            material: material_handle,
-            ..default()
-        },
-    ));
-}
